feat: separate panel for test email ui
